+++ src/pages/index.ts
@@ -1,0 +1,4 @@
+export { default as AboutPage } from './AboutPage';
+export { default as ErrorPage } from './ErrorPage';
+export { default as FormsPage } from './FormsPage';
+export { MainPage } from './MainPage';
